+++ src/components/auth/LogInForm.tsx
@@ -4,7 +4,6 @@ import * as z from "zod";
 import { auth } from "@/firebase";
 import {
   signInWithEmailAndPassword,
-  createUserWithEmailAndPassword,
   signOut,
   onAuthStateChanged,
 } from "firebase/auth";
@@ -18,22 +17,13 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { toast } from "../ui/use-toast";
+
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
 
 const FormSchema = z.object({
-  //isSeller: z.string(),
-  username: z.string().min(2, { message: "이름을 입력해주세요" }),
   email: z
     .string()
     .min(2, {
@@ -68,29 +58,10 @@ const LogInForm = () => {
   const form = useForm<z.infer<typeof FormSchema>>({
     resolver: zodResolver(FormSchema),
     defaultValues: {
-      //isSeller: "",
-      username: "",
       email: "",
       password: "",
     },
   });
-
-  const onSubmitSignUp = async (data: z.infer<typeof FormSchema>) => {
-    console.log(data);
-    try {
-      const userCredential = await createUserWithEmailAndPassword(
-        auth,
-        data.email,
-        data.password
-      );
-      console.log(userCredential);
-      localStorage.setItem("user", JSON.stringify(userCredential));
-    } catch (error) {
-      console.log(error);
-    }
-    console.log("현재 로그인한 유저 정보", auth.currentUser);
-    navigate("/");
-  };
 
   const onSubmitLogIn = async (data: z.infer<typeof FormSchema>) => {
     console.log(data);
@@ -121,55 +92,9 @@ const LogInForm = () => {
       <p className="text-2xl text-lime-400"> Today Room </p>
       <Form {...form}>
         <form
-          onSubmit={
-            to === "/logIn"
-              ? form.handleSubmit(onSubmitLogIn)
-              : form.handleSubmit(onSubmitSignUp)
-          }
+          onSubmit={form.handleSubmit(onSubmitLogIn)}
           className="w-2/3 space-y-6"
         >
-          {/* {to === "/logIn" ? null : (
-            <FormField
-              control={form.control}
-              name="isSeller"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>판매자/구매자 구분</FormLabel>
-                  <Select
-                    onValueChange={field.onChange}
-                    defaultValue={field.value}
-                  >
-                    <FormControl>
-                      <SelectTrigger>
-                        <SelectValue placeholder="판매자 혹은 구매자를 선택해주세요" />
-                      </SelectTrigger>
-                    </FormControl>
-                    <SelectContent>
-                      <SelectItem value="판매자">판매자</SelectItem>
-                      <SelectItem value="구매자">구매자</SelectItem>
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            ></FormField>
-          )} */}
-
-          {to === "/logIn" ? null : (
-            <FormField
-              control={form.control}
-              name="username"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>이름</FormLabel>
-                  <FormControl>
-                    <Input placeholder="이름을 적어주세요" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            ></FormField>
-          )}
           <FormField
             control={form.control}
             name="email"
@@ -201,16 +126,13 @@ const LogInForm = () => {
               </FormItem>
             )}
           ></FormField>
-          {to === "/logIn" ? (
-            <Button type="submit">로그인</Button>
-          ) : (
-            <Button type="submit">가입 완료</Button>
-          )}
-          {to === "/logIn" ? (
-            <Button asChild>
-              <Link to="/signUp">회원가입</Link>
-            </Button>
-          ) : null}
+
+          <Button type="submit">로그인</Button>
+
+          <Button asChild>
+            <Link to="/signUp">회원가입</Link>
+          </Button>
+
           {/* <Button onClick={logOut}>로그아웃</Button> */}
         </form>
       </Form>
