+++ src/pages/logIn/LogInPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Button } from "@/components/ui/button";
-import LogIn from "@/components/logIn/LogInForm";
+import LogIn from "@/components/auth/LogInForm";
 
 const LogInPage = () => {
   return (
