+++ src/pages/logIn/SignUpPage.tsx
@@ -1,9 +1,7 @@
 import React from "react";
 import { Button } from "@/components/ui/button";
-import LogIn from "@/components/logIn/LogInForm";
+import LogIn from "@/components/auth/LogInForm";
 import { auth } from "@/firebase";
-
-
 
 const SignUpPage = () => {
   return (
